microservices/receipt-service/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
werkzeug>=2.0
google-generativeai>=0.3.0
openai>=1.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
python-dateutil>=2.8.0
pymongo>=4.0.0
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
price-parser>=0.4.0
tiktoken>=0.9.0